% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_help_view.R
\name{stf_help_view}
\alias{stf_help_view}
\title{Function stf_help_view}
\usage{
stf_help_view()
}
\value{
This function has only side effect, it doesn't return anything.
}
\description{
This function displays STF's help page on Rstudio viewer pane.
}
\examples{
\dontrun{
stf_help_view()
}
}
\keyword{help}
\keyword{stf,}
