% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_inteiro_teor.R
\name{stf_inteiro_teor}
\alias{stf_inteiro_teor}
\title{Function stf_inteiro_teor}
\usage{
stf_inteiro_teor(url)
}
\arguments{
\item{urls}{vetor de urls dos pdfs}
}
\value{
vetor com urls dos processos agrupados de 10 em 10
}
\description{
Essa função baixa e lê os acórdãos do stf conforme url extraída via função stf_metadata
}
\examples{
Baixar e ler acórdão do STF
txt<-stf_inteiro_teor("TELEFONICA+BRASIL") 
}
\keyword{decisão}
\keyword{inteiro}
\keyword{jurisprudência,}
\keyword{stf,}
\keyword{teor,}
