% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_metadados.R
\name{stf_metadados}
\alias{stf_metadados}
\title{Function stf_metadados}
\usage{
stf_metadados(urls)
}
\arguments{
\item{urls}{endereços eletrônicos dos processos agrupados de 10 em 10}
}
\value{
Tabela com metadados dos processos
}
\description{
Essa função retorna os metadados da jurisprudencia do STF conforme a url extraída com a função stf_urls.R
}
\keyword{jurisprudência}
\keyword{stf,}
