% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_metadata.R
\name{stf_metadata}
\alias{stf_metadata}
\title{Function stf_metadata}
\usage{
stf_metadata(open_search, database = "acordaos", parts_names = TRUE)
}
\arguments{
\item{open_search}{Words to be searched}

\item{database}{Character string with one of these seven options:
"acordaos","sumulas","monocraticas","presidencia",
"sumulas_vinculantes","repercussao_geral","questoes_ordem".
Default is "acordaos".}

\item{parts_names}{Logical. If TRUE (default), it will attempt to fix 
parts prefixes.}
}
\value{
Dataframe with the metadata
}
\description{
This function returns metadada from Brazilian Supreme Court precedents
}
\keyword{metadata}
\keyword{precedents,}
\keyword{stf,}
