% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_mono_metadata.R
\name{stf_mono_metadata}
\alias{stf_mono_metadata}
\title{Returns metadada from Brazilian Supreme Court monocratic decisions}
\usage{
stf_mono_metadata(open_search, parties_names = TRUE)
}
\arguments{
\item{open_search}{Words to be searched}

\item{parties_names}{Logical. If TRUE (default), it will attempt to fix 
parties prefixes.}
}
\value{
Dataframe with the metadata
}
\description{
Returns metadada from Brazilian Supreme Court monocratic decisions
}
\keyword{metadata}
\keyword{precedents,}
\keyword{stf,}
