% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_thesaurusDF.R
\name{stf_thesaurusDF}
\alias{stf_thesaurusDF}
\title{Function stf_thesaurusDF}
\usage{
stf_thesaurusDF(letter = "", term = "")
}
\arguments{
\item{letter}{specify the letter}

\item{term}{specify the word or phrase to be searched}
}
\value{
Dataframe with the thesaurus
}
\description{
This function returns a data frame with STF thesaurus
}
\keyword{stj,}
\keyword{thesaurus}
