% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_metadata.R
\name{stf_urls}
\alias{stf_urls}
\title{Function stf_table}
\usage{
stf_urls(open_search, database = "acordaos")
}
\arguments{
\item{open_search}{Words to be searched}

\item{database}{Character string with one of these seven options:"acordaos","sumulas","monocraticas","presidencia",
"sumulas_vinculantes","repercussao_geral","questoes_ordem".
Default is "acordaos".}
}
\value{
Dataframe with the metadata
}
\description{
This function returns metadada from Brazilian Supreme Court precedents
}
\keyword{metadata}
\keyword{precedents,}
\keyword{stf,}
