% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_metadata.R
\name{stj_metadata}
\alias{stj_metadata}
\title{Function stj_metadata}
\usage{
stj_metadata(open_search, database = "acordaos", phrase = FALSE,
  start_date = "", end_date = "")
}
\arguments{
\item{open_search}{Words to be searched}

\item{database}{Character string with one of these seven options:"acordaos","sumulas",
"monocraticas',"informativo"
Default is "acordaos".}

\item{phrase}{Logical. If TRUE, it will deparse the sentence.}

\item{start_date}{Optional. Character string  in the
format: "dd/mm/yyyy" with the start of period to search for.}

\item{end_date}{Optional. Character string  in the
format: "dd/mm/yyyy" with the end of period to search for.}
}
\value{
Dataframe with the metadata
}
\description{
This function returns metadada from Brazilian Federal High Court precedents
}
\keyword{metadata}
\keyword{precedents,}
\keyword{stj,}
