% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_mono_metadata.R
\name{stj_mono_metadata}
\alias{stj_mono_metadata}
\title{Get metadata from Brazilian highest Federal Court's (STJ) monocratic decisions}
\usage{
stj_mono_metadata(open_search, phrase = TRUE)
}
\arguments{
\item{open_search}{search parameter}

\item{phrase}{logical, if TRUE, default, the search parameter will be treated as phrase.}
}
\value{
data frame with the metadata
}
\description{
Get metadata from Brazilian highest Federal Court's (STJ) monocratic decisions
}
\examples{
family_doctors<- stj_mono_metadata("m\\u00e9dico de fam\\u00edlia")
}
