% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_opinion_metadata.R
\name{stj_opinion_metadata}
\alias{stj_opinion_metadata}
\title{Gets metadada from Brazilian Federal High Court precedents}
\usage{
stj_opinion_metadata(open_search, phrase = TRUE)
}
\arguments{
\item{open_search}{Words to be searched}

\item{phrase}{logical, if TRUE, default, the search parameter will be treated as phrase.}
}
\value{
Data frame with the metadata
}
\description{
Gets metadada from Brazilian Federal High Court precedents
}
