% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_acompanhamento.R
\name{stf_acompanhamento}
\alias{stf_acompanhamento}
\title{Function stf_acompanhamento}
\usage{
stf_acompanhamento(url)
}
\arguments{
\item{url}{url dos andamentos}
}
\value{
tabela com os andamentos processuais
}
\description{
Essa função retorna a tabela com o andamento processual da jurispruência buscada com a função stf_metadados.R
}
\keyword{andamento}
\keyword{jurisprudência,}
\keyword{processual}
\keyword{stf,}
