% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_acordaos.R
\name{stf_acordaos}
\alias{stf_acordaos}
\title{Function stf_acordaos}
\usage{
stf_acordaos(df)
}
\arguments{
\item{df}{data frame with at least two columns, one with the url,
and the other, named electronic, with a logical vector informing wether the pdf is text(TRUE)
or image(FALSE).}
}
\value{
vector with the whole content of the decisions.
}
\description{
This functions reads the pdfs with the whole decisions according with the metadata
    obtained by the stf_metadata function.
}
\keyword{decision}
\keyword{inteiro}
\keyword{precedents,}
\keyword{stf,}
\keyword{teor,}
