% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_acordaos.R
\name{stf_acordaos}
\alias{stf_acordaos}
\title{Function stf_acordaos}
\usage{
stf_acordaos(df)
}
\arguments{
\item{df}{dataframe com duas colunas, uma com o vetor de urls dos pdfs e outra com  vetor lógico chamado eletronico. 
Deve-se informar quais são pdfs são eletrônicos (TRUE) e quais,
escaneados (FALSE).}
}
\value{
vetor de textos após conversão dos pdfs
}
\description{
Essa função baixa e lê os acórdãos do stf conforme url extraída via função stf_metadata
}
\keyword{decisão}
\keyword{inteiro}
\keyword{jurisprudência,}
\keyword{stf,}
\keyword{teor,}
