% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_monocratica.R
\name{stf_monocratica}
\alias{stf_monocratica}
\title{Function stf_monocratica}
\usage{
stf_monocratica(urls)
}
\arguments{
\item{pesquisa_livre}{}
}
\value{
vetor com pdfs em texto
}
\description{
Essa função baixa as decisões monocráticas do STF
}
\keyword{decisão}
\keyword{inteiro}
\keyword{jurisprudência,}
\keyword{stf,}
\keyword{teor,}
