% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_urls.R
\name{stf_urls}
\alias{stf_urls}
\title{Function stf_urls}
\usage{
stf_urls(pesquisa_livre, decisao = "acordaos")
}
\arguments{
\item{pesquisa_livre}{palavras a serem buscadas}

\item{decisao}{indicar se são acórdãos ou decisões monocráticas}
}
\value{
vetor com urls dos processos agrupados de 10 em 10
}
\description{
Essa função retorna as urls dos processos baseadas na pesquisa livre da jurisprudência do STF
}
\keyword{jurisprudência}
\keyword{stf,}
