% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_metadata.R
\name{stj_metadata}
\alias{stj_metadata}
\title{Function stj_metadata}
\usage{
stj_metadata(open_search, database = "acordaos")
}
\arguments{
\item{open_search}{Words to be searched}

\item{database}{Character string with one of these seven options:"acordaos","sumulas","monocraticas","presidencia",
"sumulas_vinculantes","repercussao_geral","questoes_ordem".
Default is "acordaos".}
}
\value{
Dataframe with the metadata
}
\description{
This function returns metadada from Brazilian Federal High Court precedents
}
\keyword{metadata}
\keyword{precedents,}
\keyword{stj,}
